Server/requirement.txt
flask
opencv-python
psycopg2-binary
scipy
sklearn
azure-cognitiveservices-vision-customvision